requirements.txt
torch~=1.11.0
torchvision~=0.12.0
numpy~=1.22.3
Pillow~=9.1.0
pandas~=1.4.2
tqdm~=4.64.0
GitPython~=3.1.27
pytorch-lightning==1.6.1
torchmetrics==0.7.0
rich==12.3.0

wandb~=0.12.16
PyYAML~=6.0
Jinja2~=3.1.2